feat(docz-example-react-native-flow): change folders
