+++ examples/react-native-flow/package.json
@@ -7,23 +7,22 @@
     "build": "docz build"
   },
   "dependencies": {
-    "docz": "^0.10.0",
-    "docz-core": "^0.10.0",
-    "prop-types": "^15.6.2",
     "react": "^16.4.2",
     "react-dom": "^16.4.2",
     "react-art": "^16.4.2",
     "react-native": "0.56.0",
     "react-native-web": "^0.8.9",
     "styled-components": "^3.4.2",
-    "docz-theme-default": "^0.10.3"
+    "docz-theme-default": "^0.10.3",
+    "prop-types": "^15.6.2"
   },
   "devDependencies": {
+    "docz": "^0.10.3",
+    "docz-core": "^0.10.3",
     "babel-jest": "23.4.2",
     "babel-plugin-module-resolver": "^3.1.1",
     "babel-plugin-react-native-web": "^0.8.9",
     "babel-preset-react-native": "^5",
-    "docz": "^0.10.3",
     "babel-plugin-react-docgen": "^2.0.0-babel7.0"
   }
 }
