+++ examples/react-native-propTypes/src/Button.js
@@ -1,6 +1,6 @@
-// @flow
 import * as React from 'react'
 import styled from 'styled-components/native'
+import t from 'prop-types'
 
 const BlueButton = styled.TouchableOpacity`
   background-color: blue;
@@ -15,15 +15,19 @@ const WhiteText = styled.Text`
   color: #fff;
 `
 
-type Props = {
-  text: string,
-  onPress: () => void
-}
-
-const Button = ({ text, onPress }: Props) => (
+const Button = ({ text, onPress }) => (
   <BlueButton onPress={onPress}>
     <WhiteText>{text}</WhiteText>
   </BlueButton>
 )
+
+Button.propTypes = {
+  text: t.string,
+  onPress: t.func,
+}
+
+Button.defaultProps = {
+  text: null,
+}
 
 export default Button
